Sort works now, changed type of value to Double, will need everyone to work off of the branch or their program might not open
