+++ app/src/main/java/com/example/team29project/Database.java
@@ -9,6 +9,7 @@ import com.google.firebase.firestore.CollectionReference;
 import com.google.firebase.firestore.EventListener;
 import com.google.firebase.firestore.FirebaseFirestore;
 import com.google.firebase.firestore.FirebaseFirestoreException;
+import com.google.firebase.firestore.Query;
 import com.google.firebase.firestore.QueryDocumentSnapshot;
 import com.google.firebase.firestore.QuerySnapshot;
 
@@ -19,11 +20,11 @@ import java.util.HashMap;
 import java.util.Objects;
 
 public class Database {
-    private final FirebaseFirestore db;
+    private FirebaseFirestore db;
 
     // Item attributes
-    private final CollectionReference itemsRef;
-    private final ArrayList<Item> itemDataList;
+    private CollectionReference itemsRef;
+    private ArrayList<Item> itemDataList;
     private ItemArrayAdapter itemAdapter;
 
     // Tag attributes
@@ -77,6 +78,7 @@ public class Database {
     }
 
     public void loadInitialItems() {
+
         // Add a snapshot listener to the Firestore reference 'itemsRef'
         itemsRef.addSnapshotListener(new EventListener<QuerySnapshot>() {
             @Override
@@ -96,7 +98,15 @@ public class Database {
                         // Retrieve various fields from the document
                         String name = doc.getId();
                         String date = doc.getString("date");
-                        double itemValue = Double.parseDouble(Objects.requireNonNull(doc.getString("value")));
+                        // Retrieve the 'value' field as an Object
+                        Object itemValueObject = doc.get("value");
+                        double itemValue = 0.0;
+                        // Check the type of 'itemValueObject' and convert it to a double
+                        if (itemValueObject instanceof Long) {
+                            itemValue = ((Long) itemValueObject).doubleValue();
+                        } else if (itemValueObject instanceof Double) {
+                            itemValue = (Double) itemValueObject;
+                        }
                         String make = doc.getString("make");
                         String model = doc.getString("model");
                         String serialNumber = doc.getString("serialNumber");
@@ -118,14 +128,15 @@ public class Database {
         // Ensure data list does not already contain item with same name
         assert (!itemDataList.contains(item));
 
-        HashMap<String, String> data = new HashMap<>();
+        HashMap<String, Object> data = new HashMap<>();
         data.put("date", item.getDate());
-        data.put("value", item.getValue().toString());
+        data.put("value", item.getValue());
         data.put("make", item.getMake());
         data.put("model", item.getModel());
         data.put("serialNumber", item.getSerialNumber());
         data.put("description", item.getDescription());
         data.put("comment", item.getComment());
+        data.put("photo", item.getPhotos());
         // Add the 'data' map to the Firestore database under a document named after the item's name.
         itemsRef
                 .document(item.getName())
@@ -205,5 +216,131 @@ public class Database {
     public boolean removeAllItems(ArrayList<Item> items) {
         return this.itemDataList.removeAll(items);
     }
+
+    public void filter(String filterBy, String data) {
+        db = FirebaseFirestore.getInstance();
+
+        Query query; // Declare a Query object
+
+        if(filterBy.equals("make")) {
+            query = db.collection("items").whereEqualTo("make", data);
+        } else if (filterBy.equals("date")) {
+            //TODO date range
+            query = db.collection("items");//need to change this query
+        } else if (filterBy.equals("description")) {
+            //TODO multiple description words or most number of words matched
+            query = db.collection("items");//need to change this query
+        } else {
+            query = db.collection("items");
+        }
+
+        // Add a snapshot listener to the Firestore query
+        query.addSnapshotListener(new EventListener<QuerySnapshot>() {
+            @Override
+            public void onEvent(@Nullable QuerySnapshot value, @Nullable FirebaseFirestoreException error) {
+                // If there's an error with the snapshot, log the error
+                if (error != null) {
+                    Log.e("Firebase", error.toString());
+                }
+
+                // If the snapshot is not null (i.e., there's data at 'itemsRef')
+                if (value != null) {
+                    // Clear the 'dataList'
+                    itemDataList.clear();
+
+                    // Loop over each document in the snapshot
+                    for (QueryDocumentSnapshot doc: value) {
+                        // Retrieve various fields from the document
+                        String name = doc.getId();
+                        String date = doc.getString("date");
+                        Number itemValue = Float.parseFloat(Objects.requireNonNull(doc.getString("value")));
+
+                        String make = doc.getString("make");
+                        String model = doc.getString("model");
+                        String serialNumber = doc.getString("serialNumber");
+                        String description = doc.getString("description");
+                        String comment = doc.getString("comment");
+
+                        // Add a new 'Item' object to 'dataList' with these fields
+                        itemDataList.add(new Item(name, date, (Double) itemValue, make, model, description, comment, serialNumber));
+                    }
+
+                    // refresh ListView and display the new data
+                    itemAdapter.notifyDataSetChanged();
+                }
+            }
+        });
+    }
+
+
+    public void sort(String sortBy, Boolean isAsc) {
+        db = FirebaseFirestore.getInstance();
+
+        // sorting the data by the sortBy field in ascending or descending order
+        Query.Direction direction = isAsc ? Query.Direction.ASCENDING : Query.Direction.DESCENDING;
+        Query query; // New Query variable
+        sortBy=sortBy.toLowerCase();
+        if (sortBy.equals("date")) {
+            query = db.collection("items").orderBy("date", direction);
+        } else if (sortBy.equals("value")) {
+            query = db.collection("items").orderBy("value", direction);
+        } else if (sortBy.equals("make")) {
+            query = db.collection("items").orderBy("make", direction);
+        } else if (sortBy.equals("model")) {
+            query = db.collection("items").orderBy("model", direction);
+        } else if (sortBy.equals("serialnumber")) {
+            query = db.collection("items").orderBy("serialnumber", direction);
+        } else if (sortBy.equals("description")) {
+            query = db.collection("items").orderBy("description", direction);
+        } else if (sortBy.equals("comment")) {
+            query = db.collection("items").orderBy("comment", direction);
+        } else {
+            query = db.collection("items");
+        }
+
+        // Add a snapshot listener to the Firestore query
+        query.addSnapshotListener(new EventListener<QuerySnapshot>() {
+            @Override
+            public void onEvent(@Nullable QuerySnapshot value, @Nullable FirebaseFirestoreException error) {
+                // If there's an error with the snapshot, log the error
+                if (error != null) {
+                    Log.e("Firebase", error.toString());
+                }
+                // If the snapshot is not null (i.e., there's data at 'itemsRef')
+                if (value != null) {
+                    // Clear the 'dataList'
+                    itemDataList.clear();
+                    //TODO figure out how to deal with null values
+                    // Loop over each document in the snapshot
+                    for (QueryDocumentSnapshot doc: value) {
+                        // Retrieve various fields from the document
+                        String name = doc.getId();
+                        String date = doc.getString("date");
+                        Object itemValueObject = doc.get("value");
+                        double itemValue = 0.0;
+                        if (itemValueObject instanceof Long) {
+                            itemValue = ((Long) itemValueObject).doubleValue();
+                        } else if (itemValueObject instanceof Double) {
+                            itemValue = (Double) itemValueObject;
+                        }
+                        String make = doc.getString("make");
+                        String model = doc.getString("model");
+                        String serialNumber = doc.getString("serialNumber");
+                        String description = doc.getString("description");
+                        String comment = doc.getString("comment");
+
+                        // Add a new 'Item' object to 'dataList' with these fields
+                        itemDataList.add(new Item(name, date, itemValue, make, model, description, comment, serialNumber));
+                    }
+
+                    // refresh ListView and display the new data
+                    itemAdapter.notifyDataSetChanged();
+                }
+            }
+        });
+    }
+
+
+
 
 }
